Refactor vector search helpers
